unificando routers em 1 arquivo
