+++ src/routes/routes.js
@@ -1,4 +1,6 @@
 import AlunoController from "../controllers/AlunoController.js";
+import ProfessorController from "../controllers/ProfessorController.js";
+
 
 const routerAluno = (app) => {
   app.get("/alunos", AlunoController.exibirTodos);
@@ -8,5 +10,14 @@ const routerAluno = (app) => {
   app.patch("/alunos/:matricula", AlunoController.atualiza);
 };
 
+const RouterProfessor=(app)=>{
+  app.get('/professor',ProfessorController.exibeProf);
+  app.get('/professor/:id',ProfessorController.exibeum);
+  app.post('/professor',ProfessorController.cadastrar);
+  app.delete('/professor/:id',ProfessorController.excluir);
+  app.patch('/professor/:id',ProfessorController.atualizar);
 
-export default routerAluno;
+}
+
+
+export {routerAluno,RouterProfessor};
